requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.14
django-bitfield==2.0.1
django-crispy-forms==1.9.0
django-csp==3.6
django-dynamic-formsets==0.0.8
django-feature-policy==3.3.0
django-filter==2.3.0
geographiclib==1.50
geopy==1.21.0
googlemaps==4.4.1
gunicorn==20.0.4
idna==2.9
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
mock==4.0.2
mysqlclient==1.4.6
numpy
pandas
prettytable==0.7.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pyUFbr==0.1.0
pyproj
requests==2.23.0
scikit-learn==0.23.1
scipy
Shapely
six==1.14.0
sqlparse==0.3.1
tqdm==4.48.0
unidecode==1.1.1
xlrd==1.2.0
cython
